+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Reject duplicate optional arguments in `rflx` CLI (eng/recordflux/RecordFlux#1342)
+
 ## [0.12.0] - 2023-08-22
 
 ### Added
@@ -354,6 +360,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2019-05-14
 
+[Unreleased]: https://github.com/AdaCore/RecordFlux/compare/v0.12.0...HEAD
 [0.12.0]: https://github.com/AdaCore/RecordFlux/compare/v0.11.1...v0.12.0
 [0.11.1]: https://github.com/AdaCore/RecordFlux/compare/v0.11.0...v0.11.1
 [0.11.0]: https://github.com/AdaCore/RecordFlux/compare/v0.10.0...v0.11.0